requirements.txt
dnspython==2.4.2
requests==2.31.0
flask==2.3.3
flask-cors==4.0.0